Allow Pedalboard to compile against JUCE 6.15+ on Red Hat. (#121)
* Redhat compilation of pedalboard

Problem

This is a solution for https://github.com/spotify/pedalboard/issues/120
When building on RedHat, and possibly other linux platforms, the
following error is seen during the C++ compilation

error: call of overloaded ‘convertFixedToFloat
    (float*&, const int*, float&, long long int&)’ is ambiguous

Solution

Conditionally, based on the preprocessor flag LINUX, cast the argument
numSamples to size_t

Result

Compilation of the C++ code will succeed on RedHat

* Update pedalboard/io/ReadableAudioFile.h


* Modifying fix so that the numSamples varaibles is always cast
to an int meaning it will be future proof and more explicit

+++ pedalboard/io/ReadableAudioFile.h
@@ -281,7 +281,7 @@ public:
         for (int c = 0; c < numChannels; c++) {
           juce::FloatVectorOperations::convertFixedToFloat(
               channelPointers[c], (const int *)channelPointers[c], scaleFactor,
-              numSamples);
+              static_cast<int>(numSamples));
         }
       }
     }
